Separate the v3 library from the v2 library
