--- include/yboardv2.h
@@ -1,161 +1,0 @@
-#ifndef YBOARDV2_H
-#define YBOARDV2_H
-
-#include <stdint.h>
-
-#include "yboard_base.h"
-
-class YBoardV2 : public YBoardBase {
-  public:
-    YBoardV2();
-    virtual ~YBoardV2();
-
-    void setup() override;
-
-    BoardType get_type() override;
-
-    ////////////////////////////// Switches/Buttons ///////////////////////////////
-    /*
-     *  This function returns the state of a switch.
-     *  The switch_idx is an integer between 1 and 2, representing the number of the
-     * target switch (for example, 1 corresponds to switch 1 on the board). The bool
-     * return type means that this function returns a boolean value (true or false).
-     *  True corresponds to the switch being on, and false corresponds to the switch
-     * being off.
-     */
-    bool get_switch(uint8_t switch_idx) override;
-
-    /*
-     *  This function returns the state of a button.
-     *  The button_idx is an integer between 1 and 3, representing the number of the
-     * target button (for example, 1 corresponds to button 1 on the board). The bool
-     * return type means that this function returns a boolean value (true or false).
-     *  True corresponds to the button being pressed, and false corresponds to the
-     * button being released.
-     */
-    bool get_button(uint8_t button_idx) override;
-
-    int get_knob();
-
-    ////////////////////////////// Speaker/Tones //////////////////////////////////
-
-    /*
-     *  This function plays a tone on the speaker.
-     *  The freq is an integer between 31 and 1174, representing the frequency of
-     * the tone in Hz. The duration is an integer representing the duration of the
-     * tone in milliseconds. The void return type means that this function does not
-     * return a value.
-     */
-    void play_note(unsigned int freq, unsigned long duration);
-
-  protected:
-    const int switch1_pin = 9;
-    const int switch2_pin = 10;
-    const int button1_pin = 11;
-    const int button2_pin = 12;
-    const int button3_pin = 13;
-    const int tone_pin = 33;
-
-  private:
-    void setup_switches();
-    void setup_buttons();
-    void setup_timer();
-
-    hw_timer_t *interrupt_timer = NULL;
-};
-
-/*
- *  These constant values are used to specify the frequency of a note to be
- * played on the speaker.
- */
-#define NOTE_B0 31
-#define NOTE_C1 33
-#define NOTE_CS1 35
-#define NOTE_D1 37
-#define NOTE_DS1 39
-#define NOTE_E1 41
-#define NOTE_F1 44
-#define NOTE_FS1 46
-#define NOTE_G1 49
-#define NOTE_GS1 52
-#define NOTE_A1 55
-#define NOTE_AS1 58
-#define NOTE_B1 62
-#define NOTE_C2 65
-#define NOTE_CS2 69
-#define NOTE_D2 73
-#define NOTE_DS2 78
-#define NOTE_E2 82
-#define NOTE_F2 87
-#define NOTE_FS2 93
-#define NOTE_G2 98
-#define NOTE_GS2 104
-#define NOTE_A2 110
-#define NOTE_AS2 117
-#define NOTE_B2 123
-#define NOTE_C3 131
-#define NOTE_CS3 139
-#define NOTE_D3 147
-#define NOTE_DS3 156
-#define NOTE_E3 165
-#define NOTE_F3 175
-#define NOTE_FS3 185
-#define NOTE_G3 196
-#define NOTE_GS3 208
-#define NOTE_A3 220
-#define NOTE_AS3 233
-#define NOTE_B3 247
-#define NOTE_C4 262
-#define NOTE_CS4 277
-#define NOTE_D4 294
-#define NOTE_DS4 311
-#define NOTE_E4 330
-#define NOTE_F4 349
-#define NOTE_FS4 370
-#define NOTE_G4 392
-#define NOTE_GS4 415
-#define NOTE_A4 440
-#define NOTE_AS4 466
-#define NOTE_B4 494
-#define NOTE_C5 523
-#define NOTE_CS5 554
-#define NOTE_D5 587
-#define NOTE_DS5 622
-#define NOTE_E5 659
-#define NOTE_F5 698
-#define NOTE_FS5 740
-#define NOTE_G5 784
-#define NOTE_GS5 831
-#define NOTE_A5 880
-#define NOTE_AS5 932
-#define NOTE_B5 988
-#define NOTE_C6 1047
-#define NOTE_CS6 1109
-#define NOTE_D6 1175
-#define NOTE_DS6 1245
-#define NOTE_E6 1319
-#define NOTE_F6 1397
-#define NOTE_FS6 1480
-#define NOTE_G6 1568
-#define NOTE_GS6 1661
-#define NOTE_A6 1760
-#define NOTE_AS6 1865
-#define NOTE_B6 1976
-#define NOTE_C7 2093
-#define NOTE_CS7 2217
-#define NOTE_D7 2349
-#define NOTE_DS7 2489
-#define NOTE_E7 2637
-#define NOTE_F7 2794
-#define NOTE_FS7 2960
-#define NOTE_G7 3136
-#define NOTE_GS7 3322
-#define NOTE_A7 3520
-#define NOTE_AS7 3729
-#define NOTE_B7 3951
-#define NOTE_C8 4186
-#define NOTE_CS8 4435
-#define NOTE_D8 4699
-#define NOTE_DS8 4978
-
-#endif /* YBOARDV2_H */
