+++ library.json
@@ -1,5 +1,5 @@
 {
-    "name": "Y-Board",
+    "name": "Y-Board v3",
     "version": "1.0.0",
     "description": "A library for interacting with version 2 and 3 of the Y-Board hardware",
     "keywords": "education, y-board",
